Update Dispose Pattern

+++ D_LOH/Program.cs
@@ -6,7 +6,7 @@ namespace LOHDemo
     {
         static void Main(string[] args)
         {
-            byte[] largeArray = new byte[100 * 1024];  // 100 KB
+            float[] largeArray = new float[100 * 1024];  // 400 KB
 
             Console.WriteLine("Large object allocated on the LOH.");
             Console.ReadKey();
